feat: locate the substrings included in message
This is the initial logic for determining which
substrings listed in 'array_of_sub_strings' are located
in the provided 'message' argument, and which
substrings are not.

The hash output will document the substring as the key name,
and the value will either be a 1 or 0. If the value is 1,
this means that the substring is included in the message
at least once, otherwise the value will remain as 0.

The next commit will include uniq count logic that helps
determine the number of times each substring is included
in the provided 'message' argument.

The hash output will also be modified so that it only presents
substrings that were found in the provided 'message' at
least once.